+++ Changelog.md
@@ -1,8 +1,13 @@
 # Changelog
 
-** 10/21/2015 **
+**10/21/2015**
 - Created using JBoss EAP=6.4.0.GA
 
-** 10/10/2017 **
+**10/10/2017**
 - Added Jboss 6.4.6 patch
 - added a new image layer for setting proper docker variables onto AUFS
+
+**05/28/2019**
+- Upgrading to 7.3.0
+- Removed support for activemq, supporting only Artemis (previously known as HornetQ)
+- removed support for startup scripts (don't even remember why I thought that was a good idea)